+++ __tests__/cases/server/server.js
@@ -25,7 +25,8 @@ const CONFIG_FIELDS = [
 	'styles',
 	'session',
 	'auth',
-	'errorHandler',
+	'error',
+	// 'errorHandler',
 	'webpack',
 	'moduleDev',
 	'bundlePathRelative',
@@ -64,7 +65,7 @@ describe('Server instance', () => {
 			styles,
 			session,
 			auth,
-			errorHandler,
+			error,
 			webpack,
 			locale,
 		} = server.getConfig();
@@ -85,7 +86,7 @@ describe('Server instance', () => {
 		expect(session).to.be.an('function');
 		expect(new session()).to.be.an.instanceOf(Session);
 		expect(auth).to.be.an('function');
-		expect(errorHandler).to.be.an('function');
+		expect(error).to.have.all.keys(['handler']);
 		expect(webpack).to.be.an('object');
 		expect(locale).to.have.all.keys(['default', 'accepted']);
 		expect(locale.default).to.be.equal('en-US');
@@ -125,6 +126,9 @@ describe('Server instance', () => {
 				accepted: ['en-US'],
 			},
 			logging: false,
+			error: {
+				layout: Layout,
+			},
 		});
 		expect(server._config).to.have.all.keys(CONFIG_FIELDS);
 		const {
@@ -141,7 +145,7 @@ describe('Server instance', () => {
 			styles,
 			session,
 			auth,
-			errorHandler,
+			error,
 			webpack,
 			locale,
 			logging,
@@ -165,7 +169,9 @@ describe('Server instance', () => {
 		expect(session).to.be.an('function');
 		expect(new session()).to.be.an.instanceOf(Session);
 		expect(auth).to.be.an('function');
-		expect(errorHandler).to.be.an('function');
+		expect(error).to.have.all.keys(['handler', 'layout']);
+		expect(error.handler).to.be.a('function');
+		expect(new error.layout()).to.be.an.instanceOf(Layout);
 		expect(webpack).to.be.an('object');
 		expect(locale).to.have.all.keys(['default', 'accepted']);
 		expect(locale.default).to.be.equal('cs-CZ');