+++ src/server/middleware/error.js
@@ -4,7 +4,7 @@ import HttpError from 'http-smart-error';
  * @param {import('../').default} server
  */
 export default (server) => (err, req, res, next) => {
-	const { dev, errorHandler } = server.getConfig();
+	const { dev, error } = server.getConfig();
 	if (!(err instanceof HttpError)) {
 		if (typeof err === 'string') {
 			// eslint-disable-next-line no-param-reassign
@@ -42,11 +42,12 @@ export default (server) => (err, req, res, next) => {
 				// eslint-disable-next-line no-underscore-dangle
 				version: server._version,
 			},
+			layout: error.layout,
 		});
 	};
-	if (typeof errorHandler !== 'function') {
+	if (!error || typeof error.handler !== 'function') {
 		render();
 		return;
 	}
-	errorHandler(err, req, res, () => render());
+	error.handler(err, req, res, () => render());
 };